removing need for requirejs config by internalizing all of the configuration needs to code that will keep track of the require depedency graph

+++ lib/util/require/optimize.coffee
@@ -5,45 +5,67 @@ requirejs = require 'requirejs'
 
 logger =    require '../logger'
 
+_ = require 'lodash'
+
+requireRegister = require './register'
+
 class Optimizer
 
   constructor: ->
     almondInPath  = path.join __dirname, "..", "assets", "almond.js"
     @almondText = fs.readFileSync almondInPath, "ascii"
 
-  optimize: (config) =>
+  optimize: (config, fileName) =>
     return unless config.optimize
-
     return if @alreadyRunning # hack right now
     @alreadyRunning = true
 
-    @setupConfig(config)
+    files = if fileName then requireRegister.treeBasesForFile(fileName) else requireRegister.treeBases()
+    numFiles = files.length
+    numProcessed = 0
+    done = (almondOutPath) =>
+      if ++numProcessed >= numFiles
+        @alreadyRunning = false
+        logger.info "Cleaning up..."
+        fs.unlink almondOutPath
+        logger.info "Requirejs optimization complete."
 
-    almondOutPath = path.join @config.baseUrl, "almond.js"
-    fs.writeFile almondOutPath, @almondText, 'ascii', (err) =>
-      return logger.error "Cannot write Almond, #{err}" if err?
-      logger.info "Beginning requirejs optimization"
-      try
-        requirejs.optimize @config, (buildResponse) =>
-          logger.success "Requirejs optimization complete.  The compiled file is ready for use.", true
-          @done(almondOutPath)
-      catch err
-        logger.error err
-        @done(almondOutPath)
+    if files?.length > 0
+      baseUrl = path.join config.watch.compiledDir, config.compilers.javascript.directory
+      almondOutPath = path.join baseUrl, "almond.js"
+      fs.writeFile almondOutPath, @almondText, 'ascii', (err) =>
+        return logger.error "Cannot write Almond, #{err}" if err?
+        for file in files
+          runConfig = @setupConfig(config, file, baseUrl)
+          logger.info "Beginning requirejs optimization for module [[#{runConfig.include[0]}]]"
+          try
+            requirejs.optimize runConfig, (buildResponse) =>
+              logger.success "The compiled file [[#{runConfig.out}]] is ready for use.", true
+              done(almondOutPath)
+          catch err
+            logger.error err
+            done(almondOutPath)
 
+  setupConfig: (config, file, baseUrl) =>
+    runConfig = _.extend({}, config.require.optimize)
+    runConfig.baseUrl = baseUrl unless runConfig.baseUrl
 
-  setupConfig: (config) =>
-    unless @config?
-      @config = config.require.optimize
-      @config.baseUrl = path.join config.watch.compiledDir, config.compilers.javascript.directory
-      @config.out = path.join @config.baseUrl, config.require.optimize.out
-      @config.include = [config.require.optimize.name]
-      @config.insertRequire = [config.require.optimize.name]
-      @config.wrap = true
-      @config.name = 'almond'
+    name = file.replace(runConfig.baseUrl + path.sep, '').replace('.js', '')
 
-  done: (almondOutPath) =>
-    @alreadyRunning = false
-    fs.unlink almondOutPath
+    paths = {}
+    for alias, configPath of requireRegister.configPaths(file)
+      paths[alias] = configPath.replace(runConfig.baseUrl + path.sep, '').replace('.js', '')
+
+    runConfig.paths = paths          unless runConfig.paths
+    runConfig.include = [name]       unless runConfig.include
+    runConfig.insertRequire = [name] unless runConfig.insertRequire
+    runConfig.wrap = true            unless runConfig.wrap
+    runConfig.name = 'almond'        unless runConfig.name
+    runConfig.out = if runConfig.out
+      path.join runConfig.baseUrl, runConfig.out
+    else
+      path.join runConfig.baseUrl, name + "-built.js"
+
+    runConfig
 
 exports.optimize = new Optimizer().optimize
